refactor: 통계 엔티티 수정 , 스케쥴링 예외처리 , 통계 날짜 범위 지정

+++ src/main/java/com/fairing/fairplay/statistics/repository/hourlystats/HourlyStatsCustomRepositoryImpl.java
@@ -36,7 +36,7 @@ public class HourlyStatsCustomRepositoryImpl implements HourlyStatsCustomReposit
                         .eventId(t.get(r.event.eventId))
                         .statDate(targetDate)
                         .hour(t.get(r.createdAt.hour()))
-                        .reservations(t.get(r.count()).intValue())
+                        .reservations(t.get(r.count()).longValue())
                         .createdAt(LocalDateTime.now())
                         .build())
                 .toList();
